apps/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "baantlo-backend"
version = "0.1.0"
description = "FastAPI backend for Baant Lo"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.114.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.4.0",
  "python-multipart>=0.0.9",
  "email-validator>=2.1.0.post1",
  "sqlalchemy>=2.0.32",
  "psycopg[binary]>=3.2.1",
  "alembic>=1.13.2",
  "celery>=5.4.0",
  "redis>=5.0.7",
  "minio>=7.2.10",
  "pyjwt[crypto]>=2.9.0",
  "passlib[bcrypt]>=1.7.4",
  "bcrypt==3.2.2",
  "google-auth>=2.34.0",
  "requests>=2.32.3",
  "httpx>=0.27.0",
  "twilio>=9.0.5",
  "psutil>=5.9.0",
  "uvloop>=0.19.0",  # High-performance event loop
  "httptools>=0.6.1",  # Fast HTTP parser
  "websockets>=12.0",  # WebSocket support
]

[tool.uvicorn]
factory = false
host = "0.0.0.0"
port = 8000
reload = true

[tool.pyright]
pythonVersion = "3.12"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"


[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
  "httpx>=0.27.0",
  "faker>=24.0.0",  # For generating fake data in development
]


[tool.setuptools]
packages = { find = {} }

[tool.setuptools.package-data]
app = ["**/*"]
